requirements.txt
numpy
sympy==1.10.1
antlr4-python3-runtime==4.7.2
openai
ivy-core
saytex
click
pickle5